+++ src/entropy/mod.rs
@@ -1,9 +1,11 @@
 //! Empirical entropy of discrete byte distributions.
 
+mod collision;
 mod hartley;
 mod renyi;
 mod shannon;
 
+pub use collision::{collision_entropy, collision_entropy_distribution};
 pub use hartley::{hartley_entropy, hartley_entropy_distribution};
 pub use renyi::{InvalidRenyiOrder, renyi_entropy, renyi_entropy_distribution};
 pub use shannon::{shannon, shannon_distribution};